# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartRest AIoT API",
        "description": "RESTful API for smart mattress monitoring system supporting patients, doctors, customers, and admins.",
        "contact": {
            "name": "SmartRest Support",
            "email": "support@smartrest.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "SmartRest API Server"
        }
    ],
    "paths": {
...
        "/messages": {
            "post": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Send a new message",
                "description": "Send a message to a recipient (patient to doctor, doctor to patient, customer to support)",
                "operationId": "sendMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_id",
                                    "body",
                                    "type"
                                ],
                                "properties": {
                                    "recipient_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Sleep Report Discussion"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "I've reviewed your latest sleep report and have some recommendations."
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "alert",
                                            "chat",
                                            "promo"
                                        ],
                                        "example": "chat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message sent successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not authorized to message this recipient",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You can only message your assigned doctors"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipient not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/thread/{conversationId}": {
            "get": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Get conversation thread",
                "description": "Retrieve message thread between the current user and another user or from a specific message",
                "operationId": "getMessageThread",
                "parameters": [
                    {
                        "name": "conversationId",
                        "in": "path",
                        "description": "User ID or message ID (prefixed with 'msg_')",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message thread",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message_id": {
                                                        "type": "string",
                                                        "example": "msg_12345"
                                                    },
                                                    "sender_id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "recipient_id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "body": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "alert",
                                                            "chat",
                                                            "promo"
                                                        ]
                                                    },
                                                    "is_read": {
                                                        "type": "boolean"
                                                    },
                                                    "sent_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Thread not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/notifications": {
            "get": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Get user notifications",
                "description": "Retrieve unread notifications and alerts for the current user",
                "operationId": "getUserNotifications",
                "responses": {
                    "200": {
                        "description": "User notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "notifications": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message_id": {
                                                        "type": "string"
                                                    },
                                                    "sender_id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "body": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "alert",
                                                            "chat",
                                                            "promo"
                                                        ]
                                                    },
                                                    "sent_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/notifications/{id}": {
            "put": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Mark notification as read",
                "description": "Mark a specific notification as read",
                "operationId": "markNotificationRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notification ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification acknowledged",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification acknowledged"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Not your notification"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}